implement creation of data and test pass
