+++ README.md
@@ -46,7 +46,7 @@ update orders set amount=1;
 
 insert mock data into the DB:
 ```
-node scripts/bin/create_from_csv.js
+node scripts/bin/create
 ```
 
 make sure that the files are compiled:
@@ -62,4 +62,20 @@ marketplace-server> npm run transpile
 After the scripts are compiled, run the tests:
 ```
 marketplace-server> npm test
+```
+
+### To run docker tests:
+
+you need to have a stellar account with funds and create a `.env` file locally with the following content:
+```
+STELLAR_CHANNEL_SEEDS=SXXX
+STELLAR_BASE_SEED=SXXX
+STELLAR_ADDRESS=GXXX
+```
+
+Then run the following commands:
+```
+make build  # build typescript
+make up  # start all services
+make test-system-docker  # run tests
 ```
