+++ config/internal.default.json
@@ -15,6 +15,8 @@
 		"synchronize": true,
 		"logging": false
 	},
+	"payment_service": "http://localhost:5000",
+	"payment_complete_callback": "http://localhost:3001/v1/internal/payments",
 	"jwt": {
 		"private_keys": {
 			"rs512_0": {
