+++ docker-compose.yaml
@@ -2,8 +2,8 @@ version: "3"
 
 services:
 
-  marketplace-service:
-    image: kinecosystem/marketplace-service
+  marketplace-server:
+    image: kinecosystem/marketplace-server
     ports:
       - 3000
     links:
@@ -19,8 +19,8 @@ services:
       APP_HOST: '0.0.0.0'
       APP_NAME: 'marketplace-server'
       APP_REDIS: redis://redis:6379/0
-      APP_PAYMENT_SERVICE: http://payment-service
-      APP_PAYMENT_COMPLETE_CALLBACK: http://marketplace-internal/v1/internal/payments
+      APP_PAYMENT_SERVICE: http://payment-service:3000
+      APP_PAYMENT_COMPLETE_CALLBACK: http://marketplace-internal:3000/v1/internal/payments
       APP_DB_TYPE: postgres
       APP_DB_USERNAME: user
       APP_DB_PASSWORD: pass
@@ -29,12 +29,13 @@ services:
       APP_DB_DATABASE: ecosystem
 
   marketplace-internal:
-    image: kinecosystem/marketplace-service
-    cmd: npm run start-internal
+    image: kinecosystem/marketplace-server
+    command: npm run start-internal
     ports:
       - 3000
     links:
       - redis
+      - payment-service
       - postgres
     environment:
       <<: *app_env_vars
@@ -47,6 +48,6 @@ services:
     ports:
       - 5432
     environment:
-      - POSTGRES_USER: user
-      - POSTGRES_PASSWORD: pass
-      - POSTGRES_DB: ecosystem
+      POSTGRES_USER: user
+      POSTGRES_PASSWORD: pass
+      POSTGRES_DB: ecosystem
