+++ scripts/src/internal/routes.ts
@@ -6,6 +6,8 @@ import {
 	paymentFailed as paymentFailedService,
 } from "./services";
 
+import { statusHandler } from "./middleware";
+
 export const paymentComplete = async function(req: Request, res: Response) {
 	await paymentCompleteService(req.body as CompletedPayment, req.logger);
 	res.status(200).send({ status: "ok" });
@@ -23,4 +25,5 @@ export function createRoutes(app: Express, pathPrefix?: string) {
 		.post("/failed-payments", paymentFailed);
 
 	app.use("/v1/internal/", router);
+	app.get("/status", statusHandler);
 }
