+++ scripts/src/public/config.ts
@@ -2,8 +2,6 @@ import { Config as BaseConfig, init as baseInit, getConfig as baseGetConfig } fr
 
 export interface Config extends BaseConfig {
 	assets_base: string;
-	payment_service: string;
-	payment_complete_callback: string;
 }
 
 export function getConfig(): Config {
@@ -20,14 +18,6 @@ function init(): void {
 	path += "default";
 
 	baseInit(`${ path }.json`);
-
-	const config = getConfig();
-	if (process.env.APP_PAYMENT_SERVICE) {
-		config.payment_service = process.env.APP_PAYMENT_SERVICE!;
-	}
-	if (process.env.APP_PAYMENT_COMPLETE_CALLBACK) {
-		config.payment_complete_callback = process.env.APP_PAYMENT_COMPLETE_CALLBACK!;
-	}
 }
 
 init();
